move library seeding logic to groove storage and encapsulate all namespace to RGT, update documentation
