+++ js/grooveStorage.js
@@ -360,6 +360,79 @@ export function purgeDanglingPatterns(activeNames) {
 }
 
 /**
+ * Checks if the library needs seeding and triggers the Welcome prompt.
+ */
+export function checkLibrarySeed() {
+  const LIBRARY_SEED_KEY = "rgt_library_seeded";
+  if (localStorage.getItem(LIBRARY_SEED_KEY)) return;
+
+  // Kill any background "Saved" toasts before showing the interactive Welcome notice
+  import("./ui/sliders.js").then((m) => m.clearNotice());
+
+  fetch("./defaultGrooves.json")
+    .then((res) => res.json())
+    .then((data) => {
+      const noticeEl = document.getElementById("uiNotice");
+      if (!noticeEl) return;
+
+      noticeEl.innerHTML = `
+        <div style="margin-bottom: 12px; font-weight: 600;">
+          Welcome: would you like to use the default grooves with actual drum sounds?
+        </div>
+        <div style="display: flex; gap: 8px; justify-content: center;">
+          <button id="seed-yes" style="background: var(--accent); color: white; min-height: 36px; padding: 0 16px;">Yes</button>
+          <button id="seed-no" style="min-height: 36px; padding: 0 16px;">No</button>
+        </div>
+      `;
+
+      noticeEl.classList.remove("hidden");
+      noticeEl.classList.add("interactive");
+      gsap.fromTo(
+        noticeEl,
+        { y: -20, opacity: 0 },
+        { y: 5, opacity: 1, duration: 0.5, ease: "back.out(1.7)" }
+      );
+
+      const libNames = Object.keys(data.library).join("\n");
+      const finalizeChoice = () => {
+        localStorage.setItem(LIBRARY_SEED_KEY, "true");
+        gsap.to(noticeEl, {
+          y: -20,
+          opacity: 0,
+          duration: 0.3,
+          onComplete: () => {
+            noticeEl.classList.add("hidden");
+            noticeEl.classList.remove("interactive");
+            if (localStorage.getItem("grooveEditorState") === "list") {
+              document.dispatchEvent(
+                new CustomEvent("metronome:ownerChanged", {
+                  detail: { owner: null },
+                })
+              );
+            }
+          },
+        });
+      };
+
+      document.getElementById("seed-yes").onclick = () => {
+        ingestLibrary(data.library, false);
+        localStorage.setItem("userGrooveNames", libNames);
+        const groovesEl = document.getElementById("grooves");
+        if (groovesEl) groovesEl.value = libNames;
+        finalizeChoice();
+      };
+
+      document.getElementById("seed-no").onclick = () => {
+        localStorage.setItem("userGrooveNames", libNames);
+        const groovesEl = document.getElementById("grooves");
+        if (groovesEl) groovesEl.value = libNames;
+        finalizeChoice();
+      };
+    })
+    .catch((err) => debugLog("state", "Library fetch skipped or failed", err));
+}
+
+/**
  * Wipes all pattern data (used by restoreDefaults).
  */
 export function clearAllGroovePatterns() {
