+++ js/main.js
@@ -9,7 +9,7 @@ import * as uiController from "./uiController.js";
 import { initDarkMode } from "./ui/theme.js";
 import * as controls from "./ui/controls.js";
 import { initModeTabs, initSimplePanelControls } from "./ui/panels.js";
-import { debugLog } from "./debug.js";
+import { debugLog, DEBUG } from "./debug.js";
 import { Profiler } from "./profiler.js";
 import { initWakeLock } from "./ui/wakeLock.js";
 import { initAdvancedMode } from "./ui/advancedMode.js";
@@ -18,82 +18,6 @@ import { initGrooveEditor, updatePlayhead } from "./ui/grooveEditor.js";
 import * as grooveStorage from "./grooveStorage.js";
 import { loadDrumSamples } from "./sampleLoader.js";
 import { ensureAudio } from "./audioProfiles.js";
-
-// Expose explicitly (redundant but safe)
-window.Profiler = Profiler;
-/**
- * Checks if the library needs seeding and triggers the Welcome prompt.
- * Globalized for re-trigger after preview discard/save.
- */
-window.checkLibrarySeed = () => {
-  const LIBRARY_SEED_KEY = "rgt_library_seeded";
-  if (localStorage.getItem(LIBRARY_SEED_KEY)) return;
-
-  // Kill any background "Saved" toasts before showing the interactive Welcome notice
-  import("./ui/sliders.js").then((m) => m.clearNotice());
-
-  fetch("./defaultGrooves.json")
-    .then((res) => res.json())
-    .then((data) => {
-      const noticeEl = document.getElementById("uiNotice");
-      if (!noticeEl) return;
-
-      noticeEl.innerHTML = `
-        <div style="margin-bottom: 12px; font-weight: 600;">
-          Welcome: would you like to use the default grooves with actual drum sounds?
-        </div>
-        <div style="display: flex; gap: 8px; justify-content: center;">
-          <button id="seed-yes" style="background: var(--accent); color: white; min-height: 36px; padding: 0 16px;">Yes</button>
-          <button id="seed-no" style="min-height: 36px; padding: 0 16px;">No</button>
-        </div>
-      `;
-
-      noticeEl.classList.remove("hidden");
-      noticeEl.classList.add("interactive");
-      gsap.fromTo(
-        noticeEl,
-        { y: -20, opacity: 0 },
-        { y: 5, opacity: 1, duration: 0.5, ease: "back.out(1.7)" }
-      );
-
-      const libNames = Object.keys(data.library).join("\n");
-      const finalizeChoice = () => {
-        localStorage.setItem(LIBRARY_SEED_KEY, "true");
-        gsap.to(noticeEl, {
-          y: -20,
-          opacity: 0,
-          duration: 0.3,
-          onComplete: () => {
-            noticeEl.classList.add("hidden");
-            noticeEl.classList.remove("interactive");
-            if (localStorage.getItem("grooveEditorState") === "list") {
-              document.dispatchEvent(
-                new CustomEvent("metronome:ownerChanged", {
-                  detail: { owner: null },
-                })
-              );
-            }
-          },
-        });
-      };
-
-      document.getElementById("seed-yes").onclick = () => {
-        grooveStorage.ingestLibrary(data.library, false);
-        localStorage.setItem("userGrooveNames", libNames);
-        const groovesEl = document.getElementById("grooves");
-        if (groovesEl) groovesEl.value = libNames;
-        finalizeChoice();
-      };
-
-      document.getElementById("seed-no").onclick = () => {
-        localStorage.setItem("userGrooveNames", libNames);
-        const groovesEl = document.getElementById("grooves");
-        if (groovesEl) groovesEl.value = libNames;
-        finalizeChoice();
-      };
-    })
-    .catch((err) => debugLog("state", "Library fetch skipped or failed"));
-};
 
 // set early to fetch from commits.json
 let appVersion;
@@ -266,7 +190,7 @@ if (document.readyState === "loading") {
   const hasSharedGroove = uiController.checkDeepLinks();
 
   // 2. Only check for library seeding if NO shared groove is being previewed
-  if (!hasSharedGroove) window.checkLibrarySeed();
+  if (!hasSharedGroove) grooveStorage.checkLibrarySeed();
 
   uiController.initMuteControl(); // Sync mute state
   loadDrumSamples(); // Load audio samples before initializing related UI
@@ -417,3 +341,27 @@ window.addEventListener("keydown", unlockAudio);
 window.addEventListener("hashchange", () => {
   uiController.checkDeepLinks();
 });
+
+// === Developer Console Bridge ===
+// Consolidates tools into a single namespaced object
+if (typeof window !== "undefined") {
+  window.RGT = Object.freeze({
+    // Logic Cores
+    metronome: metronome,
+    session: sessionEngine,
+    scheduler: patternScheduler,
+    simple: simpleMetronome,
+
+    // Developer Tools
+    debug: DEBUG,
+    profiler: Profiler,
+
+    // Utilities & Storage
+    storage: grooveStorage,
+    utils: utils,
+  });
+
+  console.info(
+    "🛠️ RGT Bridge active. Access via window.RGT (e.g., RGT.debug.audio = true)"
+  );
+}
